WebSockets/tests/submissions/Opera/interfaces/WebSocket/constants/004.html

<!doctype html>
<title>WebSockets: getting constants on prototype and object</title>
<script src=/resources/testharness.js></script>
<script src=/resources/testharnessreport.js></script>
<script src=../../../../constants.js></script>
<div id=log></div>
<script>
test(function(){
  var ws = new WebSocket(SCHEME_AND_DOMAIN+'/');
  var constants = ['CONNECTING', 'OPEN', 'CLOSING', 'CLOSED'];
  for (var i = 0; i < constants.length; ++i) {
    assert_equals(WebSocket.prototype[constants[i]], i, 'WebSocket.prototype.'+constants[i]);
    assert_equals(ws[constants[i]], i, 'ws.'+constants[i]);
  }
});
end();
</script>
